+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
  - don't allow new requests when *is ending* or *is closed*
  - write response resolving
  - wrong commands are rejected immediately
+ - handle failed commands 
  
 ### Added:
  - client can emit events (`close`)